@@ -32,13 +32,13 @@ maintainers = [
 name = "llama-index-cli"
 packages = [{include = "llama_index/"}]
 readme = "README.md"
-version = "0.3.1"
+version = "0.4.0"
 
 [tool.poetry.dependencies]
-python = ">=3.8.1,<4.0"
-llama-index-core = "^0.11.0"
-llama-index-embeddings-openai = "^0.2.0"
-llama-index-llms-openai = "^0.2.0"
+python = ">=3.9,<4.0"
+llama-index-core = "^0.12.0"
+llama-index-embeddings-openai = "^0.3.0"
+llama-index-llms-openai = "^0.3.0"
 
 [tool.poetry.group.dev.dependencies]
 black = {extras = ["jupyter"], version = "<=23.9.1,>=23.7.0"}
