@@ -28,12 +28,12 @@ exclude = ["**/BUILD"]
 license = "MIT"
 name = "llama-index-agent-openai-legacy"
 readme = "README.md"
-version = "0.2.0"
+version = "0.3.1"
 
 [tool.poetry.dependencies]
-python = ">=3.8.1,<4.0"
-llama-index-llms-openai = "^0.2.0"
-llama-index-core = "^0.11.0"
+python = ">=3.9,<4.0"
+llama-index-llms-openai = "^0.3.0"
+llama-index-core = "^0.12.0"
 
 [tool.poetry.group.dev.dependencies]
 ipython = "8.10.0"
