@@ -28,10 +28,10 @@ exclude = ["**/BUILD"]
 license = "MIT"
 name = "llama-index-callbacks-aim"
 readme = "README.md"
-version = "0.2.0"
+version = "0.3.0"
 
 [tool.poetry.dependencies]
-python = ">=3.8.1,<4.0"
+python = ">=3.9,<3.10"
 llama-index-core = "^0.10.1"
 aim = "^3.17.5"
 jwt = "^1.3.1"
