@@ -406,7 +406,9 @@ async def test_workflow_task_raises():
             raise ValueError("The step raised an error!")
 
     workflow = DummyWorkflow()
-    with pytest.raises(ValueError, match="The step raised an error!"):
+    with pytest.raises(
+        WorkflowRuntimeError, match="Error in step 'step': The step raised an error!"
+    ):
         await workflow.run()
 
 
@@ -418,7 +420,9 @@ async def test_workflow_task_raises_step():
             raise ValueError("The step raised an error!")
 
     workflow = DummyWorkflow()
-    with pytest.raises(ValueError, match="The step raised an error!"):
+    with pytest.raises(
+        WorkflowRuntimeError, match="Error in step 'step': The step raised an error!"
+    ):
         handler = workflow.run(stepwise=True)
         await handler.run_step()
 
