@@ -245,9 +245,45 @@ class ZhipuAI(FunctionCallingLLM):
             stream=False,
             tools=kwargs.get("tools", None),
             tool_choice=kwargs.get("tool_choice", None),
+            stop=kwargs.get("stop", None),
             timeout=self.timeout,
             extra_body=self.model_kwargs,
         )
+        tool_calls = raw_response.choices[0].message.tool_calls or []
+        return ChatResponse(
+            message=ChatMessage(
+                content=raw_response.choices[0].message.content,
+                role=raw_response.choices[0].message.role,
+                additional_kwargs={"tool_calls": tool_calls},
+            ),
+            raw=raw_response,
+        )
+
+    @llm_chat_callback()
+    async def achat(
+        self, messages: Sequence[ChatMessage], **kwargs: Any
+    ) -> ChatResponseAsyncGen:
+        messages_dict = self._convert_to_llm_messages(messages)
+        raw_response = self._client.chat.asyncCompletions.create(
+            model=self.model,
+            messages=messages_dict,
+            tools=kwargs.get("tools", None),
+            tool_choice=kwargs.get("tool_choice", None),
+            stop=kwargs.get("stop", None),
+            timeout=self.timeout,
+            extra_body=self.model_kwargs,
+        )
+        task_id = raw_response.id
+        task_status = raw_response.task_status
+        get_count = 0
+        while task_status not in [SUCCESS, FAILED] and get_count < self.timeout:
+            task_result = self._client.chat.asyncCompletions.retrieve_completion_result(
+                task_id
+            )
+            raw_response = task_result
+            task_status = raw_response.task_status
+            get_count += 1
+            await asyncio.sleep(1)
         tool_calls = raw_response.choices[0].message.tool_calls or []
         return ChatResponse(
             message=ChatMessage(
@@ -271,6 +307,7 @@ class ZhipuAI(FunctionCallingLLM):
                 stream=True,
                 tools=kwargs.get("tools", None),
                 tool_choice=kwargs.get("tool_choice", None),
+                stop=kwargs.get("stop", None),
                 timeout=self.timeout,
                 extra_body=self.model_kwargs,
             )
@@ -307,6 +344,7 @@ class ZhipuAI(FunctionCallingLLM):
                 stream=True,
                 tools=kwargs.get("tools", None),
                 tool_choice=kwargs.get("tool_choice", None),
+                stop=kwargs.get("stop", None),
                 timeout=self.timeout,
                 extra_body=self.model_kwargs,
             )
@@ -330,40 +368,6 @@ class ZhipuAI(FunctionCallingLLM):
                 )
 
         return gen()
-
-    @llm_chat_callback()
-    async def achat(
-        self, messages: Sequence[ChatMessage], **kwargs: Any
-    ) -> ChatResponseAsyncGen:
-        messages_dict = self._convert_to_llm_messages(messages)
-        raw_response = self._client.chat.asyncCompletions.create(
-            model=self.model,
-            messages=messages_dict,
-            tools=kwargs.get("tools", None),
-            tool_choice=kwargs.get("tool_choice", None),
-            timeout=self.timeout,
-            extra_body=self.model_kwargs,
-        )
-        task_id = raw_response.id
-        task_status = raw_response.task_status
-        get_count = 0
-        while task_status not in [SUCCESS, FAILED] and get_count < 40:
-            task_result = self._client.chat.asyncCompletions.retrieve_completion_result(
-                task_id
-            )
-            raw_response = task_result
-            task_status = raw_response.task_status
-            get_count += 1
-            await asyncio.sleep(1)
-        tool_calls = raw_response.choices[0].message.tool_calls or []
-        return ChatResponse(
-            message=ChatMessage(
-                content=raw_response.choices[0].message.content,
-                role=raw_response.choices[0].message.role,
-                additional_kwargs={"tool_calls": tool_calls},
-            ),
-            raw=raw_response,
-        )
 
     @llm_completion_callback()
     def complete(
